Invert dependency of behavior only/ophys apis
