+++ allensdk/internal/api/behavior_lims_api.py
@@ -1,11 +1,17 @@
-
-from ..core import lims_utilities
-from . import PostgresQueryMixin
+import numpy as np
 import pandas as pd
 
 from allensdk.api.cache import memoize
-from allensdk.internal.api import PostgresQueryMixin
+from allensdk.brain_observatory.behavior.stimulus_processing import get_stimulus_presentations, get_stimulus_templates, get_stimulus_metadata
+from allensdk.brain_observatory.behavior.metadata_processing import get_task_parameters
+from allensdk.brain_observatory.behavior.running_processing import get_running_df
+from allensdk.brain_observatory.behavior.rewards_processing import get_rewards
+from allensdk.brain_observatory.behavior.trials_processing import get_trials
+from allensdk.brain_observatory.running_speed import RunningSpeed
 from allensdk.brain_observatory.behavior.trials_processing import get_extended_trials
+from allensdk.internal.core import lims_utilities
+from allensdk.internal.api import PostgresQueryMixin
+from allensdk.internal.api import PostgresQueryMixin
 from allensdk.internal.core.lims_utilities import safe_system_path
 
 
@@ -21,7 +27,7 @@ class BehaviorLimsApi(PostgresQueryMixin):
         as session_uuid in the pickle returned by behavior_stimulus_file
         """
         self.behavior_experiment_id = behavior_experiment_id
-        super().__init__()
+        PostgresQueryMixin.__init__(self)
 
     def get_behavior_experiment_id(self):
         return self.behavior_experiment_id
@@ -62,3 +68,96 @@ class BehaviorLimsApi(PostgresQueryMixin):
         return cls(
             behavior_experiment_id=cls.foraging_id_to_behavior_session_id(foraging_id),
         )
+
+    @memoize
+    def get_stimulus_timestamps(self):
+        # We don't have a sync file, so we have to get vsync times from the pickle file
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        vsyncs = data["items"]["behavior"]["intervalsms"]
+        return np.hstack((0, vsyncs)).cumsum() / 1000.0  # cumulative time
+
+    @memoize
+    def get_licks(self):
+        # Get licks from pickle file instead of sync
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        stimulus_timestamps = self.get_stimulus_timestamps()
+        lick_frames = data['items']['behavior']['lick_sensors'][0]['lick_events']
+        lick_times = [stimulus_timestamps[frame] for frame in lick_frames]
+        return pd.DataFrame({'time': lick_times})
+
+    def get_stimulus_rebase_function(self):
+        # No sync times to rebase on, so just do nothing.
+        return lambda x: x
+
+    @memoize
+    def get_stimulus_frame_rate(self):
+        stimulus_timestamps = self.get_stimulus_timestamps()
+        return np.round(1 / np.mean(np.diff(stimulus_timestamps)), 0)
+
+    @memoize
+    def get_running_data_df(self):
+        stimulus_timestamps = self.get_stimulus_timestamps()
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        return get_running_df(data, stimulus_timestamps)
+
+    @memoize
+    def get_running_speed(self):
+        running_data_df = self.get_running_data_df()
+        assert running_data_df.index.name == 'timestamps'
+        return RunningSpeed(timestamps=running_data_df.index.values,
+                            values=running_data_df.speed.values)
+
+    @memoize
+    def get_stimulus_presentations(self):
+        stimulus_timestamps = self.get_stimulus_timestamps()
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        stimulus_presentations_df_pre = get_stimulus_presentations(data, stimulus_timestamps)
+        stimulus_metadata_df = get_stimulus_metadata(data)
+        idx_name = stimulus_presentations_df_pre.index.name
+        stimulus_index_df = stimulus_presentations_df_pre.reset_index().merge(stimulus_metadata_df.reset_index(), on=['image_name']).set_index(idx_name)
+        stimulus_index_df.sort_index(inplace=True)
+        stimulus_index_df = stimulus_index_df[['image_set', 'image_index', 'start_time']].rename(columns={'start_time': 'timestamps'})
+        stimulus_index_df.set_index('timestamps', inplace=True, drop=True)
+        stimulus_presentations_df = stimulus_presentations_df_pre.merge(stimulus_index_df, left_on='start_time', right_index=True, how='left')
+        assert len(stimulus_presentations_df_pre) == len(stimulus_presentations_df)
+
+        return stimulus_presentations_df[sorted(stimulus_presentations_df.columns)]
+
+    @memoize
+    def get_stimulus_templates(self):
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        return get_stimulus_templates(data)
+
+    @memoize
+    def get_rewards(self):
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        rebase_function = self.get_stimulus_rebase_function()
+        return get_rewards(data, rebase_function)
+
+    @memoize
+    def get_task_parameters(self):
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        return get_task_parameters(data)
+
+    @memoize
+    def get_trials(self):
+
+        licks = self.get_licks()
+        behavior_stimulus_file = self.get_behavior_stimulus_file()
+        data = pd.read_pickle(behavior_stimulus_file)
+        rewards = self.get_rewards()
+        rebase_function = self.get_stimulus_rebase_function()
+        trial_df = get_trials(data, licks, rewards, rebase_function)
+
+        return trial_df
+
+if __name__=="__main__":
+    api = BehaviorLimsApi(858098100)
+    print(api.get_trials())
